restructure cli directory for plugins
