+++ packages/ts-etl/package.json
@@ -7,8 +7,8 @@
   "scripts": {
     "test": "ava",
     "clean": "rm -rf dist/",
-    "test:integration": "ava integration/",
-    "test:unit": "ava src/test/",
+    "test:integration": "ava test/integration/",
+    "test:unit": "ava test/unit/",
     "start": "node --loader=ts-node/esm --experimental-specifier-resolution=node --nolazy -r ts-node/register/transpile-only src/index.ts",
     "lint": "npx eslint .",
     "lint:fix": "npx eslint --fix .",
