+++ packages/ts-etl/src/index.ts
@@ -2,7 +2,7 @@ import cli from './cli'
 export default cli
 
 export * from './cli'
-export * from './crossref'
+export * from './plugins/crossref'
 export * from './types'
 
 await cli.parseAsync()
